+++ src/mqt/predictor/reg/__init__.py
@@ -7,9 +7,15 @@ See README.md or go to https://github.com/cda-tum/mqt-predictor for more informa
 from __future__ import annotations
 
 from mqt.predictor.reg.hellinger_distance_regressor import (
+    calc_device_specific_features,
     get_hellinger_model_path,
     hellinger_distance,
     hellinger_model_available,
 )
 
-__all__ = ["get_hellinger_model_path", "hellinger_distance", "hellinger_model_available"]
+__all__ = [
+    "calc_device_specific_features",
+    "get_hellinger_model_path",
+    "hellinger_distance",
+    "hellinger_model_available",
+]
